refactor: Split commands.py into cli/ package and service layer
- Create src/board/cli/ package with separate command files
- Extract business logic to src/board/service.py
- Add shared CLI helpers in cli/_helpers.py
- Remove monolithic commands.py (was 1143 lines)

Addresses review feedback about:
- Single responsibility violation
- Long functions (173-201 lines)
- Config loading duplication
- Mixed concerns (business logic + presentation)
